Added Flask-JWT-Extended and claims

+++ resources/item.py
@@ -1,5 +1,6 @@
+from flask_jwt_extended import jwt_required, get_jwt
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+
 from models.item import ItemModel
 
 
@@ -38,7 +39,11 @@ class Item(Resource):
 
         return item.json(), 201
 
+    @jwt_required()
     def delete(self, name):
+        claims = get_jwt()
+        if not claims['is_admin']:
+            return {'message': 'Admin privilege requires'}, 401
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
